rfx-leaf-system-config+hook

+++ foliage/src/rectangle/proc_gen.rs
@@ -20,7 +20,7 @@ fn textures() {
             for y in 0..size {
                 for x in 0..size {
                     if x > unit {
-                        let index = x + size as u32 * y;
+                        let index = x + size * y;
                         *filled_data.get_mut(index as usize).unwrap() += 1f32;
                     }
                 }
